Rename module "memories" and "storages"
