+++ camel/memories/memory_record.py
