+++ camel/storages/dict_storage/base.py
