+++ test/memory/test_chat_history_memory.py
@@ -17,10 +17,10 @@ from pathlib import Path
 
 import pytest
 
-from camel.memory import ChatHistoryMemory, MemoryRecord
-from camel.memory.context_creator import DefaultContextCreator
+from camel.memories import ChatHistoryMemory, MemoryRecord
+from camel.memories.context_creator import DefaultContextCreator
 from camel.messages import BaseMessage
-from camel.storage.dict_storage import InMemoryDictStorage, JsonStorage
+from camel.storages.dict_storage import InMemoryDictStorage, JsonStorage
 from camel.typing import ModelType, OpenAIBackendRole, RoleType
 from camel.utils.token_counting import OpenAITokenCounter
